# Conflicts:
#	installation/template/index.php
#	installation/template/js/setup.js
#	installation/template/js/template.js

+++ administrator/components/com_media/resources/scripts/store/actions.es6.js
@@ -2,29 +2,23 @@ import { api } from '../app/Api.es6';
 import * as types from './mutation-types.es6';
 import translate from '../plugins/translate.es6';
 import { notifications } from '../app/Notifications.es6';
-import * as FileSaver from '../../../../../../node_modules/file-saver/src/FileSaver';
 
-// Actions are similar to mutations, the difference being that:
-// - Instead of mutating the state, actions commit mutations.
-// - Actions can contain arbitrary asynchronous operations.
+const updateUrlPath = (path) => {
+  const currentPath = path === null ? '' : path;
+  const url = new URL(window.location.href);
 
-// TODO move to utils
-function updateUrlPath(path) {
-  if (path == null) {
-    // eslint-disable-next-line no-param-reassign
-    path = '';
-  }
-  const url = window.location.href;
-  const pattern = new RegExp('\\b(path=).*?(&|$)');
-
-  if (url.search(pattern) >= 0) {
-    // eslint-disable-next-line no-restricted-globals
-    history.pushState(null, '', url.replace(pattern, `$1${path}$2`));
+  if (url.searchParams.has('path')) {
+    window.history.pushState(null, '', url.href.replace(/\b(path=).*?(&|$)/, `$1${currentPath}$2`));
   } else {
-    // eslint-disable-next-line no-restricted-globals
-    history.pushState(null, '', `${url + (url.indexOf('?') > 0 ? '&' : '?')}path=${path}`);
+    window.history.pushState(null, '', `${url.href + (url.href.indexOf('?') > 0 ? '&' : '?')}path=${currentPath}`);
   }
-}
+};
+
+/**
+ * Actions are similar to mutations, the difference being that:
+ * Instead of mutating the state, actions commit mutations.
+ * Actions can contain arbitrary asynchronous operations.
+ */
 
 /**
  * Get contents of a directory from the api
@@ -44,7 +38,7 @@ export const getContents = (context, payload) => {
       context.commit(types.SET_IS_LOADING, false);
     })
     .catch((error) => {
-      // TODO error handling
+      // @todo error handling
       context.commit(types.SET_IS_LOADING, false);
       // eslint-disable-next-line no-console
       console.log('error', error);
@@ -64,7 +58,7 @@ export const getFullContents = (context, payload) => {
       context.commit(types.SET_IS_LOADING, false);
     })
     .catch((error) => {
-      // TODO error handling
+      // @todo error handling
       context.commit(types.SET_IS_LOADING, false);
       // eslint-disable-next-line no-console
       console.log('error', error);
@@ -81,7 +75,7 @@ export const download = (context, payload) => {
     .then((contents) => {
       const file = contents.files[0];
 
-      // Converte the base 64 encoded string to a blob
+      // Convert the base 64 encoded string to a blob
       const byteCharacters = atob(file.content);
       const byteArrays = [];
 
@@ -99,8 +93,14 @@ export const download = (context, payload) => {
         byteArrays.push(byteArray);
       }
 
-      // Open the save as file dialog
-      FileSaver.saveAs(new Blob(byteArrays, { type: file.mime_type }), file.name);
+      // Download file
+      const blobURL = URL.createObjectURL(new Blob(byteArrays, { type: file.mime_type }));
+      const a = document.createElement('a');
+      a.href = blobURL;
+      a.download = file.name;
+      document.body.appendChild(a);
+      a.click();
+      document.body.removeChild(a);
     })
     .catch((error) => {
       // eslint-disable-next-line no-console
@@ -129,6 +129,9 @@ export const toggleBrowserItemSelect = (context, payload) => {
  * @param payload object with the new folder name and its parent directory
  */
 export const createDirectory = (context, payload) => {
+  if (!api.canCreate) {
+    return;
+  }
   context.commit(types.SET_IS_LOADING, true);
   api.createDirectory(payload.name, payload.parent)
     .then((folder) => {
@@ -137,7 +140,7 @@ export const createDirectory = (context, payload) => {
       context.commit(types.SET_IS_LOADING, false);
     })
     .catch((error) => {
-      // TODO error handling
+      // @todo error handling
       context.commit(types.SET_IS_LOADING, false);
       // eslint-disable-next-line no-console
       console.log('error', error);
@@ -150,6 +153,9 @@ export const createDirectory = (context, payload) => {
  * @param payload object with the new folder name and its parent directory
  */
 export const uploadFile = (context, payload) => {
+  if (!api.canCreate) {
+    return;
+  }
   context.commit(types.SET_IS_LOADING, true);
   api.upload(payload.name, payload.parent, payload.content, payload.override || false)
     .then((file) => {
@@ -172,22 +178,29 @@ export const uploadFile = (context, payload) => {
 /**
  * Rename an item
  * @param context
- * @param payload object: the old and the new path
+ * @param payload object: the item and the new path
  */
 export const renameItem = (context, payload) => {
+  if (!api.canEdit) {
+    return;
+  }
+
+  if (typeof payload.item.canEdit !== 'undefined' && payload.item.canEdit === false) {
+    return;
+  }
   context.commit(types.SET_IS_LOADING, true);
-  api.rename(payload.path, payload.newPath)
+  api.rename(payload.item.path, payload.newPath)
     .then((item) => {
       context.commit(types.RENAME_SUCCESS, {
         item,
-        oldPath: payload.path,
+        oldPath: payload.item.path,
         newName: payload.newName,
       });
       context.commit(types.HIDE_RENAME_MODAL);
       context.commit(types.SET_IS_LOADING, false);
     })
     .catch((error) => {
-      // TODO error handling
+      // @todo error handling
       context.commit(types.SET_IS_LOADING, false);
       // eslint-disable-next-line no-console
       console.log('error', error);
@@ -199,11 +212,17 @@ export const renameItem = (context, payload) => {
  * @param context
  */
 export const deleteSelectedItems = (context) => {
+  if (!api.canDelete) {
+    return;
+  }
   context.commit(types.SET_IS_LOADING, true);
   // Get the selected items from the store
   const { selectedItems } = context.state;
   if (selectedItems.length > 0) {
     selectedItems.forEach((item) => {
+      if (typeof item.canDelete !== 'undefined' && item.canDelete === false) {
+        return;
+      }
       api.delete(item.path)
         .then(() => {
           context.commit(types.DELETE_SUCCESS, item);
@@ -211,13 +230,13 @@ export const deleteSelectedItems = (context) => {
           context.commit(types.SET_IS_LOADING, false);
         })
         .catch((error) => {
-          // TODO error handling
+          // @todo error handling
           context.commit(types.SET_IS_LOADING, false);
           // eslint-disable-next-line no-console
           console.log('error', error);
         });
     });
   } else {
-    // TODO notify the user that he has to select at least one item
+    // @todo notify the user that he has to select at least one item
   }
 };
